chore(hpu): Add dedicated clippy_hpu_backend entry in makefile
Fix all generated clippy warning/errors

+++ backends/tfhe-hpu-backend/src/interface/cmd.rs
@@ -75,7 +75,7 @@ impl HpuCmd {
             .collect::<Vec<_>>();
         let imm_op = imm
             .iter()
-            .map(|var| Immediat::from_cst(*var as u128))
+            .map(|var| Immediat::from_cst(*var))
             .collect::<Vec<_>>();
 
         let op = IOp::new(opcode, dst_op, src_op, imm_op);
